+++ includes/admin/class-updater.php
@@ -1,0 +1,55 @@
+<?php
+/**
+ * WP-Autoplugin Updater class.
+ *
+ * @package WP-Autoplugin
+ */
+
+namespace WP_Autoplugin\Admin;
+
+use WP_Autoplugin\GitHub_Updater;
+
+if ( ! defined( 'ABSPATH' ) ) {
+	exit;
+}
+
+/**
+ * Class that handles the GitHub updater.
+ */
+class Updater {
+
+	/**
+	 * Constructor.
+	 */
+	public function __construct() {
+		add_action( 'init', [ $this, 'github_updater_init' ] );
+	}
+
+	/**
+	 * Initialize the GitHub updater.
+	 *
+	 * @return void
+	 */
+	public function github_updater_init() {
+		if ( ! is_admin() ) {
+			return;
+		}
+
+		$config = [
+			'slug'               => plugin_basename( WP_AUTOPLUGIN_DIR . 'wp-autoplugin.php' ),
+			'proper_folder_name' => dirname( plugin_basename( WP_AUTOPLUGIN_DIR . 'wp-autoplugin.php' ) ),
+			'api_url'            => 'https://api.github.com/repos/WP-Autoplugin/wp-autoplugin',
+			'raw_url'            => 'https://raw.githubusercontent.com/WP-Autoplugin/wp-autoplugin/main/',
+			'github_url'         => 'https://github.com/WP-Autoplugin/wp-autoplugin',
+			'zip_url'            => 'https://github.com/WP-Autoplugin/wp-autoplugin/archive/refs/heads/main.zip',
+			'requires'           => '6.0',
+			'tested'             => '6.6.2',
+			'description'        => esc_html__( 'A plugin that generates other plugins on-demand using AI.', 'wp-autoplugin' ),
+			'homepage'           => 'https://github.com/WP-Autoplugin/wp-autoplugin',
+			'version'            => WP_AUTOPLUGIN_VERSION,
+		];
+
+		// Instantiate the updater class.
+		new GitHub_Updater( $config );
+	}
+}
