+++ views/page-extend-theme.php
@@ -85,5 +85,5 @@ $theme_data = wp_get_theme( $theme_slug );
 		</form>
 		<div id="extended-theme-plugin-message" class="autoplugin-message"></div>
 	</div>
-	<?php $this->output_admin_footer(); ?>
+	<?php $this->admin->output_admin_footer(); ?>
 </div>
