+++ views/page-generate-plugin.php
@@ -112,5 +112,5 @@ if ( ! defined( 'ABSPATH' ) ) {
 		</form>
 		<div id="create-plugin-message" class="autoplugin-message"></div>
 	</div>
-<?php $this->output_admin_footer(); ?>
+<?php $this->admin->output_admin_footer(); ?>
 </div>
